+++ src/lmu_sm_interface/InternalsPlugin.hpp
@@ -16,7 +16,11 @@
 
 #include "PluginObjects.hpp"     // base class for plugin objects to derive from
 #include <cmath>                 // for sqrt()
+#ifdef _WIN32
 #include <windows.h>             // for HWND
+#else
+typedef void* HWND;
+#endif
 // rF2 and plugins must agree on structure packing, so set it explicitly here.
 // Whatever the current packing is will be restored at the end of this include
 // with another #pragma.