+++ src/lmu_sm_interface/PluginObjects.hpp
@@ -1,0 +1,9 @@
+#ifndef _PLUGIN_OBJECTS_HPP_
+#define _PLUGIN_OBJECTS_HPP_
+
+class PluginObject {
+public:
+    virtual ~PluginObject() {}
+};
+
+#endif