+++ src/lmu_sm_interface/SharedMemoryInterface.hpp
@@ -1,5 +1,6 @@
 #pragma once
 #include "InternalsPlugin.hpp"
+#include <optional>
 
 /*
 * Usage example: